rename directory to web-component
