+++ packages/widget/web-component/README.md
