+++ packages/widget/web-component/package.json
